writing routes

+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+  "encoding/json"
   "fmt"
 	"log"
 	"net/http"
@@ -33,37 +34,61 @@ func watch(filePath string, cb func()) error {
 	return nil
 }
 
-func nodemon(program string) {
-	cmd := exec.Command("node", program)
+var cmd *exec.Cmd
+func restartStaticService() {
+  var err error
 
-	err := cmd.Start()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-  watch(program, func() {
-    log.Printf("%s changed", program)
-
+  if (cmd != nil && cmd.Process != nil) {
     err = cmd.Process.Kill()
     if err != nil {
       log.Fatal(err)
     }
-    log.Println("Process killed")
+  }
 
-    cmd = exec.Command("node", program)
-    err = cmd.Start()
-    if err != nil {
-      log.Fatal(err)
-    }
-    log.Println("Process restarted")
-  })
+	cmd = exec.Command("node", "node-app/app.js")
+	err = cmd.Start()
+
+	if err != nil {
+		log.Fatal(err)
+	}
+}
+
+type route struct {
+  Rethod string `json:"method"`
+  Rath string `json:"path"`
+  ReturnValue string `json:"returnValue"`
+}
+
+var routes = []route{
+  route{"get", "/", "0"},
+}
+func writeRouteModule() {
+  routesJson, err := json.Marshal(routes)
+  if err != nil {
+    log.Fatal(err)
+  }
+  os.WriteFile("node-app/routes.json", routesJson, 0666)
 }
 
 func main() {
-	go nodemon("node-app/app.js")
+  // read the current routes file into memory
+
+  // if routes file does not exist, write out default
+  writeRouteModule()
+
+  // start static service
+  restartStaticService()
+
+  // when route file changes, update routes in memory
+
+  // when route file changes, restart static service
+  go watch("node-app/routes.json", restartStaticService)
 
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
+    // update route file, restart static service
+    // proxy request to static service
     fmt.Fprintf(w, "Hello from Go")
   })
+
 	log.Fatal(http.ListenAndServe(":3001", nil))
 }
